rag-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI/ML
openai==1.10.0
anthropic==0.8.1
pinecone-client==3.0.0
sentence-transformers==2.3.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
neo4j==5.16.0

# HTTP
httpx==0.26.0
requests==2.31.0

# Utilities
python-multipart==0.0.6
pyyaml==6.0.1
